add FileChooser and split gtk test with internal and out test
internal test in gtk package and out test in gtk_test package
